Move player slot components to PlayerSlot folder for better organization
