+++ client/src/components/waiting-room/PlayerSlot/EmptyPlayerSlot.stories.tsx
