+++ client/src/components/waiting-room/PlayerSlot/EmptyPlayerSlot.tsx
