+++ client/src/components/waiting-room/PlayerSlot/FilledPlayerSlot.stories.tsx
