+++ client/src/components/waiting-room/PlayerSlot/FilledPlayerSlot.tsx
