+++ client/src/components/waiting-room/PlayerSlot/PlayerSlot.tsx
